s5e8825-common: Patch out libswdap device check
This library checks for the device name before starting the effect (eg. SM-A536B),
and if the device name doesn't match, it starts in a 'crippled' mode and will refuse
to read dax-default.xml properly, so no effects (except volume-based ones) will
be applied.

The following message will appear in logcat to indicate this:

07-03 23:30:28.390   750   787 D DlbDap2Process: There's no mix matrix
07-03 23:30:28.390   750   787 D DlbDap2Process: DAP output mode set to 1 with output channel mask 0x3 and null mix matrix.

Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>