s5e8825-common: Patch audio HAL to fix low call volume
The way it works currently is that modem call volume is handled by the
"ABOX Speech Volume" mixer control, which in our prebuilt HAL gets set
through the audio HAL seemingly querying sec-ril, which returns 4.

Most users complain that this is too low, so this patch hardcodes the
value 7 in the function diamondvoice_set_volume, which is the highest
value that has any effect on the volume.

Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>